Bifurcated file into two scripts
